17 lmdberror too many open files (#25)
`Repo.env` will leave open file descriptors (and hence throw an `lmdb.Error` exception if too many are open). To avoid this, you must call `Repo.close()`, or preferably use `Repo` as a contextmanager so that we can be assured the env will be properly closed regardless of errors.

daggerml_cli.api now uses the Repo as a contextmanager always.