51 make everything hashable again (#55)
* added `Commit.dag_name` and dags aren't added to dags map until commit

Dags are hashable now. To achieve this, we move the dags map insert to commit, and add a `dag_name` property to the `Commit` object. This is a good idea anyway as there is currently no way to really tell which dag was added in a commit (i.e. there's no connection between a dag's commit message and the dag itself -- this fixes that).